refactor(cache-heal): extract inline healScript to hooks/cache-heal.mjs
